# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: CEX REST API
  version: 1.0.0
  description: REST API for the cryptocurrency exchange platform
servers:
  - url: /api/v1
    description: API v1
paths:
...
  /orders:
    post:
      tags:
        - Trading
      summary: Create order
      x-implementation-status: implemented
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            example:
              symbol: "BTC-USD"
              side: "buy"
              type: "limit"
              price: "100.00"
              quantity: "1.0"
              time_in_force: "GTC"
      responses:
        '200':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateOrderResponse'
              example:
                order_id: "2f12a5c6-6b91-4a4d-87db-2f91a7f2d6aa"
                status: "accepted"
                created_at: "2026-02-04T12:00:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags:
        - Trading
      summary: List orders
      x-implementation-status: implemented
      security:
        - bearerAuth: []
      parameters:
        - name: symbol
          in: query
          schema:
            type: string
          description: Filter by trading pair
        - name: status
          in: query
          schema:
            type: string
          description: Filter by order status
        - name: from
          in: query
          schema:
            type: string
            format: date-time
          description: Start time filter
        - name: to
          in: query
          schema:
            type: string
            format: date-time
          description: End time filter
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Orders list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrdersListResponse'
              example:
                orders:
                  - order_id: "2f12a5c6-6b91-4a4d-87db-2f91a7f2d6aa"
                    client_order_id: "client-123"
                    symbol: "BTC-USD"
                    side: "buy"
                    type: "limit"
                    price: "100.00"
                    quantity: "1.0"
                    filled_quantity: "0"
                    status: "open"
                    time_in_force: "GTC"
                    created_at: "2026-02-04T12:00:00Z"
                    updated_at: "2026-02-04T12:00:00Z"
                next_cursor: "2026-02-04T12:00:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /orders/{order_id}:
    delete:
      tags:
        - Trading
      summary: Cancel order
      x-implementation-status: implemented
      security:
        - bearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
          description: Order ID
      responses:
        '200':
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelOrderResponse'
              example:
                order_id: "2f12a5c6-6b91-4a4d-87db-2f91a7f2d6aa"
                status: "cancelled"
                updated_at: "2026-02-04T12:05:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags:
        - Trading
      summary: Get order
      x-implementation-status: implemented
      security:
        - bearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
          description: Order ID
      responses:
        '200':
          description: Order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
              example:
                order_id: "2f12a5c6-6b91-4a4d-87db-2f91a7f2d6aa"
                client_order_id: "client-123"
                symbol: "BTC-USD"
                side: "buy"
                type: "limit"
                price: "100.00"
                quantity: "1.0"
                filled_quantity: "0"
                status: "open"
                time_in_force: "GTC"
                created_at: "2026-02-04T12:00:00Z"
                updated_at: "2026-02-04T12:00:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for user authentication
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key identifier for HMAC authentication. Must be used together with X-API-Signature 
        and X-API-Timestamp headers. The signature is computed using HMAC-SHA256 over the 
        request method, path, query string, request body, and timestamp using the API key secret.
    apiSignature:
      type: apiKey
      in: header
      name: X-API-Signature
      description: HMAC-SHA256 signature of the request (required with apiKeyAuth)
    apiTimestamp:
      type: apiKey
      in: header
      name: X-API-Timestamp
      description: Unix timestamp in seconds for request signing (required with apiKeyAuth)
  parameters:
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
      description: Maximum number of items to return
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
      description: Pagination cursor (opaque)
  schemas:
...
    CreateOrderRequest:
      type: object
      required:
        - symbol
        - side
        - type
        - quantity
      properties:
        symbol:
          type: string
          example: BTC-USD
        side:
          type: string
          enum:
            - buy
            - sell
        type:
          type: string
          enum:
            - limit
            - market
        price:
          type: string
          description: Required for limit orders
          example: "45000.00"
        quantity:
          type: string
          example: "0.10"
        time_in_force:
          type: string
          enum:
            - GTC
            - IOC
            - FOK
          default: GTC
    CreateOrderResponse:
      type: object
      properties:
        order_id:
          type: string
        status:
          type: string
          example: accepted
        created_at:
          type: string
          format: date-time
    CancelOrderResponse:
      type: object
      properties:
        order_id:
          type: string
        status:
          type: string
          example: cancelled
        updated_at:
          type: string
          format: date-time
    OrderResponse:
      type: object
      properties:
        order_id:
          type: string
        client_order_id:
          type: string
        symbol:
          type: string
        side:
          type: string
          example: buy
        type:
          type: string
          example: limit
        price:
          type: string
          nullable: true
          example: "100.00"
        quantity:
          type: string
          example: "1.0"
        filled_quantity:
          type: string
          example: "0"
        status:
          type: string
          example: open
        time_in_force:
          type: string
          example: GTC
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    OrdersListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: '#/components/schemas/OrderResponse'
        next_cursor:
          type: string
          nullable: true
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          enum:
            - INVALID_REQUEST
            - UNAUTHORIZED
            - FORBIDDEN
            - RATE_LIMITED
            - INSUFFICIENT_BALANCE
            - ORDER_NOT_FOUND
            - SYMBOL_HALTED
            - INTERNAL_ERROR
            - NOT_FOUND
          description: |
            - INVALID_REQUEST: Invalid request parameters
            - UNAUTHORIZED: Authentication required or invalid credentials
            - FORBIDDEN: Insufficient permissions
            - RATE_LIMITED: Rate limit exceeded
            - INSUFFICIENT_BALANCE: Insufficient balance for operation
            - ORDER_NOT_FOUND: Order not found
            - SYMBOL_HALTED: Trading halted for symbol
            - INTERNAL_ERROR: Internal server error
            - NOT_FOUND: Resource not found
        message:
          type: string
          description: Human-readable error message
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: INVALID_REQUEST
            message: invalid payload
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: UNAUTHORIZED
            message: invalid credentials
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: FORBIDDEN
            message: insufficient permissions
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: INTERNAL_ERROR
            message: internal error